backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.26.0
cloudinary==1.38.0
slowapi==0.1.9
firebase-admin==6.4.0
stripe==7.11.0
bcrypt==4.1.2
